homework_4/requirements.txt
cassandra-driver
flask
pandas
requests